Activity Code Refactor
