+++ src/modules/tests/activity-code/activity-code.actions.ts
@@ -1,0 +1,13 @@
+
+import { ActivityCode } from '@dvsa/mes-test-schema/categories/B';
+import { Action } from '@ngrx/store';
+
+export const SET_ACTIVITY_CODE = '[Tests] Set activity code';
+
+export class SetActivityCode implements Action {
+  readonly type = SET_ACTIVITY_CODE;
+  constructor(public payload: ActivityCode) {}
+}
+
+export type Types =
+  | SetActivityCode;
