+++ src/modules/tests/activity-code/activity-code.reducer.ts
@@ -1,0 +1,16 @@
+import { createFeatureSelector } from '@ngrx/store';
+import { ActivityCode } from '@dvsa/mes-test-schema/categories/B';
+import * as activityCodeActions from './activity-code.actions';
+
+export const initialState: ActivityCode = null;
+
+export function activityCodeReducer(state = initialState, action: activityCodeActions.Types): ActivityCode {
+  switch (action.type) {
+    case activityCodeActions.SET_ACTIVITY_CODE:
+      return action.payload;
+    default:
+      return state;
+  }
+}
+
+export const getActivityCode = createFeatureSelector<ActivityCode>('activityCode');
